src/requirements.txt
flask-cors
flask
SQLAlchemy
gunicorn
PyJWT
python-dotenv